update sqr to int

+++ sushu.py
@@ -1,6 +1,6 @@
 import math
 def prime(n):
-    for i in range(2,math.sqrt(n)):
+    for i in range(2,int(math.sqrt(n))):
         if n%i==0:
             return False
         return True
